PostView2: Committed missing files. Modified curvature fields so that the FECurvatureField stores the parameters. Added property list for math data.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@9902 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostViewLib/FEState.cpp
@@ -8,7 +8,7 @@
 void ValArray::append(int items)
 {
 	if (m_index.empty()) m_index.push_back(0);
-	m_index.push_back(m_data.size() + items);
+	m_index.push_back((int)m_data.size() + items);
 	for (int i=0; i<items; ++i) m_data.push_back(0.f);
 }
